perf metricgroup: Initial support to build sys metric table
When adding system metrics, the current code iters through the generated
table in pmu-events.c::pmu_sys_event_tables, trying to match metrics.

This "works" for currently supportly platforms, but has various
limitations:
- Cannot support testcases like PMU events or parse metrics
- Cannot support metric reuse
- Relies on metrics always having "compat" property, when strictly not
  necessary

Improve this by building a table dynamically of supported metrics for the
current platform, similar to what we have for CPU events in
pmu-events::pmu_events_map. This will allow to support the above mentioned
features.

Initial support is added to find the number of all possible metrics and
allocates table memory to hold them all, before code is added to fill
in the tables.

Signed-off-by: John Garry <john.garry@huawei.com>