perf tests: Add sys metric support for parse metric testcase
Add support for system metrics to the parse metric testcase, to ensure
that we can properly compute results for system metrics.

A fake PMU with the .id member is set to match versus the JSON event
COMPAT string.

Test JSONs are added in test_soc/sys folder, which will also be covered
in the PMU events testcase.

Signed-off-by: John Garry <john.garry@huawei.com>