scsi: libsas: Queue SMP commands as requests
Send SMP commands through the block layer so that each command gets a
unique tag associated.

Function sas_task_complete_internal() is what the LLDD calls to signal
that the CQ is complete and this calls into the SCSI midlayer. And then
sas_blk_end_sync_rq() is called when the request completes.

Signed-off-by: John Garry <john.g.garry@oracle.com>