scsi: target: iblock/configfs/core: Atomic emulation configuration support
Support configuring iblock for atomic emulation. The atomic properties are
read from the backend device request queue limits.

Configfs also has support for emulate_atomic attribute, which is default
off.

Finally vpd b0 emulation support needs to be modified to provide a full-
length table. Currently if tpu and tpws emulation is disabled, then only
the top portion of the table is filled in, avoiding the unmap properties.
However since the atomic properties lie after the unmap properties in the
table, we need to provide the full table for when atomic emulation is
enabled.

TODO: check if atomic support affects unmap properties (when unmap
emulation disabled).

Signed-off-by: John Garry <john.g.garry@oracle.com>