perf jevents: Raise exception for no definition of a arch std event
Recently Ilkka reported that the arm64-based platform AmpereOne JSONs
included a duff event which referenced an non-existent arch std event [0].

Previously in the times of jevents.c, we would raise an exception for this.

This is still invalid, so re-introduce code to raise an exception for no
definition of a referenced arch std event exists.

[0] https://lore.kernel.org/linux-perf-users/9e851e2a-26c7-ba78-cb20-be4337b2916a@oracle.com/

Signed-off-by: John Garry <john.g.garry@oracle.com>