auxdisplay: Use uts_release
Instead of using UTS_RELEASE, use uts_release, which means that we don't
need to rebuild the code just for the git head commit changing when
CONFIG_LOCALVERSION_AUTO is enabled.

For linedisp_register() in line-display.c we need to snprint a string
including uts_release as this is required for linedisp_display() ->
linedisp_display().

Do something similar for charlcd_init() in charlcd.c

Signed-off-by: John Garry <john.g.garry@oracle.com>