fs: iomap: Atomic write support
Support atomic writes by producing a single BIO with REQ_ATOMIC flag set.

We rely on the FS to guarantee extent alignment, such that an atomic write
should never straddle two or more extents.

An atomic write size is limited by this extent alignment, and so in the
generic_atomic_write_valid() call we use the extent size as the unit_max
value. We will not yet support a value less than the FS block size, so
use that size for the unit_min. It is the task of the FS to set
iomap.extent_shift per iter. It is the role of the block layer to ensure
that the BIO submitted does not exceed the bdev request_queue limits.

We will treat any sub-extent IO like sub-fsblock DIO, in that we will zero
the sub-extent when the mapping is unwritten.

Signed-off-by: John Garry <john.g.garry@oracle.com>