filemap: Change mapping_set_folio_min_order() -> mapping_set_folio_orders()
Borrowed from:

https://lore.kernel.org/linux-fsdevel/20240213093713.1753368-2-kernel@pankajraghav.com/

We will need to be able to only use a single folio order for buffered
atomic writes, so allow the folio order min and max be set.

We still have the restriction of nothing be able to support order-1
folios - it be will be required to lift this limit at some stage.

Signed-off-by: John Garry <john.g.garry@oracle.com>