block: Catch possible entries missing from hctx_flag_name[]
Refresh values BLK_MQ_F_x enum, and then re-arrange members in
hctx_flag_name[] to match that enum.

An entry for NO_SCHED_BY_DEFAULT is missing, so add one.

Finally add a BUILD_BUG_ON() call to ensure that we are not missing entries
in hctx_flag_name[].

Signed-off-by: John Garry <john.g.garry@oracle.com>